api/requirements.txt
asgiref==3.6.0
click==8.1.3
colorama==0.4.6
Django==4.1.7
django-cors-headers==3.14.0
djangorestframework==3.14.0
Flask==2.2.3
gunicorn==20.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
PyMySQL==1.0.3
pytz==2023.3
sqlparse==0.4.3
tzdata==2023.3
Werkzeug==2.2.3
